requirements.txt
youtube-dl==2020.6.16
PyNaCl==1.3.0
discord.py==1.3.3
beautifulsoup4==4.9.1
requests==2.22.0